feat: Navigation to Task detail screen working properly. User can now view deatils about a task on the Task detail screen

+++ app/src/main/java/com/edozie/reve_jc/util/Routes.kt
@@ -11,7 +11,7 @@ sealed class Screen(val route: String) {
     //    object Home : Screen("home")
     object AddTask : Screen("add_task")
     object TaskDetail : Screen("task_detail/{taskId}") {
-        fun createRoute(taskId: Long) = "task_detail/$taskId"
+        fun createRoute(taskId: Int) = "task_detail/$taskId"
     }
     object Today : Screen("today")
 }
